Inicio da logica
