+++ pokedex4/src/Routes/Coordinator.js
@@ -1,0 +1,7 @@
+export const goToPokedex = (navigate) => {
+    navigate("/pokedex")
+}
+
+export const goToDetails = (navigate) => {
+    navigate("/details")
+}
